requirements.txt
autopep8==1.5.4
paho-mqtt==1.5.1
pycodestyle==2.6.0
toml==0.10.2
